feat(sdk,docs): 단말 SDK F2 — Windows C API cimsue_c.h + c_api.cpp(공개 C++ 표면 1:1 평탄화) + 단위시험
cimsue.dll 이 C++ 클래스와 함께 C API 79 함수를 export 한다 — .NET 파사드(P/Invoke)가 보는 유일한 표면.
- cimsue_c.h: 불투명 핸들, 정수 id, 동기 cimsue_status_t(=Result::code)+cimsue_last_error(스레드별), cimsue_listener_t
  콜백 한 벌(Listener 1:1, 이벤트 스레드), POD 평탄화·(ptr,count) 배열, 열거형 값 동일, cimsue_*_default() 규약
  (문자열 NULL = C++ 기본값 유지). 조회 산출은 스레드별 스냅샷(다음 조회까지 유효), CSC 산출은 핸들의 다음 호출까지.
- c_api.cpp: 프로토콜 로직 없는 변환 층. 기동 중 start 재호출은 리스너를 건드리지 않고 즉시 실패(이벤트 스레드와의 경쟁 방지).
- 요청 token(affiliate/sendRequest/RequestResult.token) long → int64_t — Windows long 은 32비트(§6.2 규칙).
- sdk/core CMake: 코어를 OBJECT 라이브러리(cimsue_objs)+INTERFACE(cimsue_iface)로 나눠 cimsue(Linux 정적/Windows DLL)와
  단위시험이 오브젝트를 공유. Windows cimsue_test 는 DLL 이 export 하지 않는 내부 심볼을 시험하므로 오브젝트를 직접 링크
  (CIMSUE_API= override — export.h). pasf 의존 교차검증(floor_xcheck)은 Linux 전용으로 분기.
- sdk/windows 슈퍼빌드: googletest v1.14 ExternalProject(CIMSUE_BUILD_TESTS, /MD, gtest_force_shared_crt) — cimsue_test 27
  케이스 Windows 통과(c_api_test 포함).
- 문서: ue_sdk.md §6.2·§6.4·§10, sdk/windows README·dotnet README, CLAUDE.md 인덱스, 루트 CMake 의 vs14.sln 주석 정정.

+++ sdk/core/cli/main.cpp
@@ -218,7 +218,7 @@ public:
         set([&] { sds.push_back(m); });
     }
     void onRequestResult(const RequestResult& r) override {
-        std::fprintf(stderr, "[cimsue-cli] %s token=%ld → %d %s etag=%s\n", r.method.c_str(), r.token, r.code, r.reason.c_str(), r.etag.c_str());
+        std::fprintf(stderr, "[cimsue-cli] %s token=%lld → %d %s etag=%s\n", r.method.c_str(), (long long)r.token, r.code, r.reason.c_str(), r.etag.c_str());
         set([&] { results[r.token] = r; });
     }
     void onMessage(int, const std::string& from, const std::string& ct, const std::string& body) override {
@@ -238,7 +238,7 @@ public:
     std::vector<DialogInfo> dialogs;
     int granted = 0, taken = 0, denied = 0, rosters = 0;
     std::vector<SdsMessage> sds;
-    std::map<long, RequestResult> results;
+    std::map<int64_t, RequestResult> results;
 
 private:
     template <typename F> void set(F f) { { std::lock_guard<std::mutex> lk(m_); f(); } cv_.notify_all(); }
@@ -387,7 +387,7 @@ int main(int argc, char** argv) {
     s.code = ls.reg.code; s.reason = ls.reg.reason;
 
     for (auto& g : o.affiliate) {
-        long tok = eng.affiliate(acc, g, true);
+        int64_t tok = eng.affiliate(acc, g, true);
         bool got = ls.waitFor([&] { return ls.results.count(tok) > 0; }, 10);
         if (!got || ls.results[tok].code / 100 != 2)
             std::fprintf(stderr, "[cimsue-cli] affiliate %s failed (code=%d)\n", g.c_str(), got ? ls.results[tok].code : 0);
